deps/v8/test/inspector/runtime/regress-1321833-expected.txt

Regression test for crbug.com/1321833

Running test: testNumberNaN
log[
    [0] : {
        className : Number
        description : Number
        objectId : 1.1.1
        preview : {
            description : Number
            overflow : false
            properties : [
                [0] : {
                    name : [[PrimitiveValue]]
                    type : number
                    value : NaN
                }
            ]
            type : object
        }
        type : object
    }
]

Running test: testNumberInfinity
log[
    [0] : {
        className : Number
        description : Number
        objectId : 1.1.2
        preview : {
            description : Number
            overflow : false
            properties : [
                [0] : {
                    name : [[PrimitiveValue]]
                    type : number
                    value : Infinity
                }
            ]
            type : object
        }
        type : object
    }
]

Running test: testNumberMinusInfinity
log[
    [0] : {
        className : Number
        description : Number
        objectId : 1.1.3
        preview : {
            description : Number
            overflow : false
            properties : [
                [0] : {
                    name : [[PrimitiveValue]]
                    type : number
                    value : -Infinity
                }
            ]
            type : object
        }
        type : object
    }
]
